+++ ExilenceNextApp/src/interfaces/api/api-stashtab-priceditem.interface.ts
@@ -1,7 +1,7 @@
-import { IPricedItem } from './api-priced-item.interface';
+import { IApiPricedItem } from './api-priced-item.interface';
 
 export interface IApiStashTabPricedItem {
   uuid: string;
   stashTabId: string;
-  pricedItems: IPricedItem[];
+  pricedItems: IApiPricedItem[];
 }