+++ ExilenceNextApp/src/utils/snapshot.utils.ts
@@ -1,13 +1,13 @@
 import moment from 'moment';
 import { rootStore } from '..';
-import { IPricedItem } from '../interfaces/api/api-priced-item.interface';
 import { IApiSnapshot } from '../interfaces/api/api-snapshot.interface';
 import { IApiStashTabSnapshot } from '../interfaces/api/api-stash-tab-snapshot.interface';
 import { IApiStashTabPricedItem } from '../interfaces/api/api-stashtab-priceditem.interface';
 import { IChartStashTabSnapshot } from '../interfaces/chart-stash-tab-snapshot.interface';
+import { IPricedItem } from '../interfaces/priced-item.interface';
 import { IStashTab } from '../interfaces/stash.interface';
 import { Snapshot } from '../store/domains/snapshot';
-import { getRarityIdentifier, mergeItemStacks } from './item.utils';
+import { findItem, getRarityIdentifier, mergeItemStacks } from './item.utils';
 
 export const mapSnapshotToApiSnapshot = (snapshot: Snapshot, stashTabs?: IStashTab[]) => {
   const filteredLeagueTabs = stashTabs?.filter((st) =>
@@ -114,7 +114,84 @@ export const formatStashTabSnapshotsForChart = (
     .sort((n1, n2) => n1[0] - n2[0]);
 };
 
-export const filterItems = (snapshots: IApiSnapshot[]) => {
+export const diffSnapshots = (snapshot1: IApiSnapshot, snapshot2: IApiSnapshot): IPricedItem[] => {
+  const difference: IPricedItem[] = [];
+  const itemsInSnapshot1 = mergeItemStacks(snapshot1.stashTabs.flatMap((sts) => sts.pricedItems));
+  const itemsInSnapshot2 = mergeItemStacks(snapshot2.stashTabs.flatMap((sts) => sts.pricedItems));
+
+  // items that exist in snapshot 2 but not in snapshot 1 & items that exist in both snapshots but should be updated
+  const itemsToAdd = itemsInSnapshot2.filter((x) => findItem(itemsInSnapshot1, x) === undefined);
+  const itemsToUpdate = itemsInSnapshot2.filter((x) => {
+    const foundItem = findItem(itemsInSnapshot1, x);
+    return foundItem !== undefined && x.stackSize !== foundItem.stackSize;
+  });
+  itemsToUpdate.concat(itemsToAdd).map((item) => {
+    const existingItem = findItem(itemsInSnapshot1, item);
+    if (existingItem) {
+      const recentItem = Object.assign({}, item);
+      recentItem.stackSize = recentItem.stackSize - existingItem.stackSize;
+      existingItem.total = recentItem.calculated * existingItem.stackSize;
+      recentItem.total = recentItem.total - existingItem.total;
+      if (recentItem.total !== 0 && recentItem.stackSize !== 0) {
+        difference.push(recentItem);
+      }
+    } else if (item.total !== 0 && item.stackSize !== 0) {
+      difference.push(item);
+    }
+  });
+
+  // items that exist in snapshot 1 but not in snapshot 2
+  const itemsToRemove = itemsInSnapshot1.filter((x) => findItem(itemsInSnapshot2, x) === undefined);
+  itemsToRemove.map((item) => {
+    const existingItem = findItem(itemsInSnapshot2, item);
+    const recentItem = Object.assign({}, item);
+    if (recentItem.total !== 0 && recentItem.stackSize !== 0) {
+      if (existingItem !== undefined) {
+        recentItem.stackSize = existingItem.stackSize - recentItem.stackSize;
+        existingItem.total = recentItem.calculated * existingItem.stackSize;
+        recentItem.total = existingItem.total - recentItem.total;
+      } else {
+        recentItem.total = -Math.abs(recentItem.total);
+        recentItem.stackSize = -Math.abs(recentItem.stackSize);
+      }
+      difference.push(recentItem);
+    }
+  });
+  return difference;
+};
+
+export const filterItems = (items: IPricedItem[]) => {
+  const filterText = rootStore.uiStateStore.bulkSellView
+    ? rootStore.uiStateStore.bulkSellItemTableFilterText.toLowerCase()
+    : rootStore.uiStateStore.itemTableFilterText.toLowerCase();
+
+  const rarity = getRarityIdentifier(filterText);
+
+  let itemNameRegex = new RegExp('', 'i');
+  try {
+    // try/catch required because of filtering being an onChange event, example: typing only [ would lead to a SyntaxError
+    itemNameRegex = new RegExp(filterText, 'i');
+  } catch (error) {
+    console.error(error);
+  }
+
+  return mergeItemStacks(
+    items.filter(
+      (i) =>
+        (i.calculated > 0 && i.name.toLowerCase().includes(filterText)) ||
+        (i.tab &&
+          i.tab
+            .map((t) => t.name)
+            .join(', ')
+            .toLowerCase()
+            .includes(filterText)) ||
+        (i.calculated > 0 && rarity >= 0 && i.frameType === rarity) ||
+        itemNameRegex.test(i.name)
+    )
+  );
+};
+
+export const filterSnapshotItems = (snapshots: IApiSnapshot[]) => {
   if (snapshots.length === 0) {
     return [];
   }